reorganize file before finish parsing
